Extend documentation
